refactor: organize finance pages
